Added Outputs
Added six PWM servo outputs to control flight controller. Also added
pass-through function to allow user control.

+++ waypointNavigator/waypointNavigator/rx.cpp
@@ -229,7 +229,7 @@ void rx_update()
 	
 }
 
-float rx_get_throttle()
+int rx_get_throttle()
 {
 	return rxInputThrottle;
 }
